Move NFA implementation to separate file
