+++ lib/dregex/state_machine.rb
@@ -1,6 +1,7 @@
 require 'fiber'
 
 require 'dregex/ast_nodes'
+require 'dregex/state_machine/nfa'
 
 module Dregex
   class StateMachine
@@ -29,42 +30,31 @@ module Dregex
     end
 
     class Builder
-      attr_reader :state, :states, :state_namer, :dispatcher
-      attr_accessor :stay_on_current_state
+      extend Forwardable
 
-      def initialize(state_namer: nil, dispatcher: nil)
-        if state_namer.nil?
-          state_namer = _build_state_namer
-        end
+      attr_reader :dispatcher, :nfa, :state
+      attr_accessor :stay_on_current_state
+      def_delegators :nfa, :states
+
+      def initialize(dispatcher: nil, nfa: NFA.new)
         if dispatcher.nil?
           dispatcher = Dregex::AstDispatcher.new(self)
         end
-        @state_namer = state_namer
-        @state = :start
-        @states = { @state => {} }
+        @nfa = nfa
         @dispatcher = dispatcher
+        @state = nfa.create_state(:start)
         @stay_on_current_state = false
-      end
-
-      def _build_state_namer
-        i = 0
-        Enumerator.new do |yielder|
-          while(1)
-            yielder << "S#{i}"
-            i += 1
-          end
-        end
       end
 
       def build_from(ast)
         dispatcher.visit ast
-        converted_states = convert_to_state_machine
+        converted_states = nfa.convert_to_state_machine
         StateMachine.new converted_states[:start]
       end
 
       def on_sequence(node)
         Fiber.new do
-          states[state][:end] = true
+          nfa.states[state][:end] = true
         end
       end
 
@@ -90,9 +80,7 @@ module Dregex
           self.stay_on_current_state = false
           state
         else
-          new = state_namer.next
-          states[new] = Hash.new
-          new
+          nfa.create_state
         end
       end
 
@@ -107,20 +95,7 @@ module Dregex
         when AstNode::Any
           :any
         end
-        states[state][value] = to_state
-      end
-
-      def convert_to_state_machine
-        convert = states.keys.to_h do |state_name|
-          [state_name, states[state_name].dup]
-        end
-        convert.each do |state_name, state|
-          state.each do |transition, next_state_name|
-            next unless convert.has_key? next_state_name
-            state[transition] = convert[next_state_name]
-          end
-        end
-        convert
+        nfa.states[state][value] = to_state
       end
 
     end
